+++ src/jiratui/actions/tests/test_actions.py
@@ -78,8 +78,8 @@ from jiratui.widgets.work_item_info.info import WorkItemInfoContainer
 from jiratui.widgets.work_item_info.screens import EditTextContentScreen
 from jiratui.widgets.work_item_info.tabs import InfoTabbedContent
 from jiratui.widgets.work_item_subtasks.subtasks import (
-    ChildWorkItemCollapsible,
-    IssueChildWorkItemsWidget,
+    SubtaskCollapsible,
+    SubtasksWidget,
     WorkItemSubtasks,
 )
 from jiratui.widgets.work_item_worklog.screens import WorkItemWorkLogScreen, WorkLogCollapsible
@@ -187,7 +187,7 @@ def bindings() -> dict:
             get_application_key_bindings()
             .get('focus_work_item_subtasks_tab', {})
             .get('keys', [])[0],
-            IssueChildWorkItemsWidget,
+            SubtasksWidget,
         ),
     ],
 )
@@ -3957,7 +3957,7 @@ async def test_key_to_edit_jql_opening_modal_screen(
         assert isinstance(app.screen, JQLEditorScreen)
 
 
-@patch.object(IssueChildWorkItemsWidget, 'action_create_work_item_subtask')
+@patch.object(SubtasksWidget, 'action_create_work_item_subtask')
 @patch('jiratui.widgets.screen.MainScreen._search_work_items')
 @patch('jiratui.widgets.screen.MainScreen.fetch_statuses')
 @patch('jiratui.widgets.screen.MainScreen.fetch_issue_types')
@@ -3998,7 +3998,7 @@ async def test_action_create_work_item_subtask_from_subtasks_tab(
         action_create_work_item_subtask_mock.assert_called_once()
 
 
-@patch.object(ChildWorkItemCollapsible, 'action_view_work_item')
+@patch.object(SubtaskCollapsible, 'action_view_work_item')
 @patch('jiratui.widgets.screen.MainScreen._search_work_items')
 @patch('jiratui.widgets.screen.MainScreen.fetch_statuses')
 @patch('jiratui.widgets.screen.MainScreen.fetch_issue_types')
@@ -4046,7 +4046,7 @@ async def test_action_view_work_item_from_selected_subtasks_in_subtasks_tab(
         action_view_work_item_mock.assert_called_once()
 
 
-@patch.object(ChildWorkItemCollapsible, 'action_open_go_to_screen')
+@patch.object(SubtaskCollapsible, 'action_open_go_to_screen')
 @patch('jiratui.widgets.screen.MainScreen._search_work_items')
 @patch('jiratui.widgets.screen.MainScreen.fetch_statuses')
 @patch('jiratui.widgets.screen.MainScreen.fetch_issue_types')
@@ -4092,6 +4092,54 @@ async def test_action_open_go_to_screen_with_selected_subtask_from_subtasks_tab(
         await app.workers.wait_for_complete()
         # THEN
         action_open_go_to_screen_mock.assert_called_once()
+
+
+@patch.object(SubtaskCollapsible, 'action_delete_work_item')
+@patch('jiratui.widgets.screen.MainScreen._search_work_items')
+@patch('jiratui.widgets.screen.MainScreen.fetch_statuses')
+@patch('jiratui.widgets.screen.MainScreen.fetch_issue_types')
+@patch('jiratui.widgets.screen.MainScreen.fetch_projects')
+@pytest.mark.asyncio
+async def test_action_delete_work_item_with_selected_subtask_from_subtasks_tab(
+    fetch_projects_mock: AsyncMock,
+    fetch_issue_types_mock: AsyncMock,
+    fetch_statuses_mock: AsyncMock,
+    search_work_items_mock: AsyncMock,
+    action_delete_work_item_mock: Mock,
+    jira_issues,
+    bindings: dict,
+    app,
+):
+    # test the action that opens the screen to view related tasks for a selected subtask
+    # GIVEN
+    app.config.search_results_truncate_work_item_summary = 10
+    app.config.search_results_style_work_item_status = False
+    app.config.search_results_style_work_item_type = False
+    app.config.search_results_per_page = 10
+    app.config.git_repositories = None
+    app.config.jira_base_url = 'foo.bar'
+    app.config.enable_goto = True
+    search_work_items_mock.return_value = WorkItemSearchResult(
+        response=JiraIssueSearchResponse(issues=jira_issues),
+        total=1,
+        start=1,
+        end=1,
+    )
+    app.config.pre_defined_jql_expressions = None
+    async with app.run_test() as pilot:
+        await pilot.press(bindings.get('search', {}).get('keys', [])[0])
+        await app.workers.wait_for_complete()
+        app.screen.issue_child_work_items_widget.issues = WorkItemSubtasks(
+            work_item_key=jira_issues[1].key,
+            project_key=jira_issues[1].project.key,
+            issues=[jira_issues[0]],
+        )
+        await pilot.press(bindings.get('focus_work_item_subtasks_tab', {}).get('keys', [])[0])
+        await pilot.press('tab')
+        await pilot.press(bindings.get('delete_work_item', {}).get('keys', [])[0])
+        await app.workers.wait_for_complete()
+        # THEN
+        action_delete_work_item_mock.assert_called_once()
 
 
 @patch.object(Widget, 'action_page_up')