+++ src/jiratui/widgets/work_item_subtasks/tests/test_subtasks.py
@@ -5,11 +5,12 @@ import pytest
 from jiratui.api_controller.controller import APIController, APIControllerResponse
 from jiratui.models import JiraIssue, JiraIssueSearchResponse
 from jiratui.widgets.screen import MainScreen
+from jiratui.widgets.screens.confirmation import ConfirmationScreen
 from jiratui.widgets.screens.goto import GoToScreen
 from jiratui.widgets.screens.work_item_quick_view import WorkItemQuickViewScreen
 from jiratui.widgets.work_item_subtasks.subtasks import (
-    ChildWorkItemCollapsible,
-    IssueChildWorkItemsWidget,
+    SubtaskCollapsible,
+    SubtasksWidget,
     WorkItemSubtasks,
 )
 
@@ -27,7 +28,7 @@ async def test_sets_subtasks(mock_configuration, jira_issues, app):
     async with app.run_test():
         # GIVEN
         mock_configuration.jira_base_url = 'http://foo.bar'
-        widget = IssueChildWorkItemsWidget()
+        widget = SubtasksWidget()
         await app.screen.mount(widget)
         # WHEN
         widget.issues = WorkItemSubtasks(
@@ -39,7 +40,7 @@ async def test_sets_subtasks(mock_configuration, jira_issues, app):
         assert widget._work_item_key == 'WI-1'
         assert widget._work_item_project_key == 'P1'
         assert len(widget.children) == 2
-        assert isinstance(widget.children[0], ChildWorkItemCollapsible)
+        assert isinstance(widget.children[0], SubtaskCollapsible)
         assert widget.children[0]._work_item_key == 'key-1'
         assert widget.children[1]._work_item_key == 'key-2'
 
@@ -48,7 +49,7 @@ async def test_sets_subtasks(mock_configuration, jira_issues, app):
 async def test_sets_subtasks_none(jira_issues, app):
     async with app.run_test():
         # GIVEN
-        widget = IssueChildWorkItemsWidget()
+        widget = SubtasksWidget()
         await app.screen.mount(widget)
         # WHEN
         widget.issues = None
@@ -62,7 +63,7 @@ async def test_sets_subtasks_none(jira_issues, app):
 async def test_view_subtask(app):
     async with app.run_test() as pilot:
         # GIVEN
-        widget = ChildWorkItemCollapsible(work_item_key='key-1')
+        widget = SubtaskCollapsible(work_item_key='key-1')
         await app.screen.mount(widget)
         # WHEN
         await widget.action_view_work_item()
@@ -77,7 +78,7 @@ async def test_related_issues_widget_opens_quick_view_screen(mock_configuration,
     mock_configuration.jira_base_url = 'http://foo.bar'
     async with app.run_test() as pilot:
         # WHEN
-        widget = IssueChildWorkItemsWidget()
+        widget = SubtasksWidget()
         await app.screen.mount(widget)
         await app.workers.wait_for_complete()
         widget.issues = WorkItemSubtasks(
@@ -102,7 +103,7 @@ async def test_open_goto_screen_with_goto_enabled(
     app.config.enable_goto = True
     async with app.run_test() as pilot:
         # WHEN
-        widget = IssueChildWorkItemsWidget()
+        widget = SubtasksWidget()
         await app.screen.mount(widget)
         await app.workers.wait_for_complete()
         widget.issues = WorkItemSubtasks(
@@ -126,7 +127,7 @@ async def test_open_goto_screen_with_goto_disabled(
     app.config.enable_goto = False
     async with app.run_test() as pilot:
         # WHEN
-        widget = IssueChildWorkItemsWidget()
+        widget = SubtasksWidget()
         await app.screen.mount(widget)
         await app.workers.wait_for_complete()
         widget.issues = WorkItemSubtasks(
@@ -142,9 +143,9 @@ async def test_open_goto_screen_with_goto_disabled(
 
 
 @patch.object(APIController, 'get_issue')
-@patch.object(ChildWorkItemCollapsible, '_close_goto_screen')
+@patch.object(SubtaskCollapsible, '_close_goto_screen')
 @pytest.mark.asyncio
-async def test_dismiss_goto_screen_with_key(
+async def test_select_item_in_goto_screen(
     close_goto_screen_mock: Mock,
     get_issue_mock: AsyncMock,
     mock_configuration,
@@ -159,9 +160,9 @@ async def test_dismiss_goto_screen_with_key(
     app.config.enable_goto = True
     async with app.run_test() as pilot:
         # WHEN
-        widget = IssueChildWorkItemsWidget()
+        widget = SubtasksWidget()
         await app.screen.mount(widget)
-        await app.workers.wait_for_complete()
+        await pilot.pause()
         widget.issues = WorkItemSubtasks(
             work_item_key=jira_issues[0].key,
             project_key='P1',
@@ -175,3 +176,106 @@ async def test_dismiss_goto_screen_with_key(
         # THEN
         assert isinstance(app.screen, MainScreen)
         close_goto_screen_mock.assert_called_once_with('key-1')
+
+
+@patch.object(APIController, 'delete_work_item')
+@pytest.mark.asyncio
+async def test_delete_subtask(
+    delete_work_item_mock: AsyncMock, mock_configuration, jira_issues, app
+):
+    # GIVEN
+    delete_work_item_mock.return_value = APIControllerResponse(success=True)
+    async with app.run_test() as pilot:
+        mock_configuration.jira_base_url = 'http://foo.bar'
+        widget = SubtasksWidget()
+        await app.screen.mount(widget)
+        widget.issues = WorkItemSubtasks(
+            work_item_key='WI-1',
+            project_key='P1',
+            issues=jira_issues,
+        )
+        assert len(widget.query_children(SubtaskCollapsible)) == 2
+        # WHEN
+        await widget.delete_work_item(SubtaskCollapsible.WorkItemDeleted(jira_issues[0].key))
+        await pilot.pause()
+        # THEN
+        delete_work_item_mock.assert_awaited_once_with(jira_issues[0].key)
+        assert len(widget.query(SubtaskCollapsible)) == 1
+
+
+@patch.object(APIController, 'delete_work_item')
+@pytest.mark.asyncio
+async def test_delete_remaining_subtask(
+    delete_work_item_mock: AsyncMock, mock_configuration, jira_issues, app
+):
+    # GIVEN
+    delete_work_item_mock.return_value = APIControllerResponse(success=True)
+    async with app.run_test() as pilot:
+        mock_configuration.jira_base_url = 'http://foo.bar'
+        widget = SubtasksWidget()
+        await app.screen.mount(widget)
+        widget.issues = WorkItemSubtasks(
+            work_item_key='WI-1',
+            project_key='P1',
+            issues=jira_issues[:1],
+        )
+        assert len(widget.query_children(SubtaskCollapsible)) == 1
+        # WHEN
+        await widget.delete_work_item(SubtaskCollapsible.WorkItemDeleted(jira_issues[0].key))
+        await pilot.pause()
+        # THEN
+        delete_work_item_mock.assert_awaited_once_with(jira_issues[0].key)
+        assert len(widget.query(SubtaskCollapsible)) == 0
+
+
+@patch.object(APIController, 'delete_work_item')
+@pytest.mark.asyncio
+async def test_delete_subtask_fails(
+    delete_work_item_mock: AsyncMock, mock_configuration, jira_issues, app
+):
+    # GIVEN
+    delete_work_item_mock.return_value = APIControllerResponse(success=False)
+    async with app.run_test() as pilot:
+        mock_configuration.jira_base_url = 'http://foo.bar'
+        widget = SubtasksWidget()
+        await app.screen.mount(widget)
+        widget.issues = WorkItemSubtasks(
+            work_item_key='WI-1',
+            project_key='P1',
+            issues=jira_issues,
+        )
+        assert len(widget.query_children(SubtaskCollapsible)) == 2
+        # WHEN
+        await widget.delete_work_item(SubtaskCollapsible.WorkItemDeleted(jira_issues[0].key))
+        await pilot.pause()
+        # THEN
+        delete_work_item_mock.assert_awaited_once_with(jira_issues[0].key)
+        assert len(widget.query(SubtaskCollapsible)) == 2
+
+
+@pytest.mark.asyncio
+async def test_delete_request_opens_confirmation_screen(mock_configuration, jira_issues, app):
+    # GIVEN
+    async with app.run_test():
+        mock_configuration.jira_base_url = 'http://foo.bar'
+        widget = SubtaskCollapsible(work_item_key='WI-1')
+        await app.screen.mount(widget)
+        # WHEN
+        await widget.action_delete_work_item()
+        # THEN
+        assert isinstance(app.screen, ConfirmationScreen)
+
+
+@pytest.mark.asyncio
+async def test_delete_request_does_not_confirmation_screen_with_missing_key(
+    mock_configuration, jira_issues, app
+):
+    # GIVEN
+    async with app.run_test():
+        mock_configuration.jira_base_url = 'http://foo.bar'
+        widget = SubtaskCollapsible(work_item_key='')
+        await app.screen.mount(widget)
+        # WHEN
+        await widget.action_delete_work_item()
+        # THEN
+        assert isinstance(app.screen, MainScreen)